refactor: fix useAuth cannot run when parent is RSC
