+++ client/components/edit-button.tsx
@@ -1,19 +1,13 @@
 'use client'
 
 import { Button } from '@/components/ui/button'
-import { isAuthenticated } from '@/lib/auth'
+import { useAuth } from '@/lib/auth'
 import { cn } from '@/lib/utils'
 import { Pencil1Icon } from '@radix-ui/react-icons'
 import Link from 'next/link'
-import { useEffect, useState } from 'react'
 
 function EditInfoButton({ uniId }: { uniId: string }) {
-  // TODO: Fix cannot use useAuth because its parent is RSC
-  const [isAuth, setIsAuth] = useState(false)
-
-  useEffect(() => {
-    setIsAuth(isAuthenticated())
-  }, [typeof window !== 'undefined'])
+  const isAuth = useAuth()
 
   return (
     <Link
