+++ client/components/review-section/edit-delete-review.tsx
@@ -1,23 +1,11 @@
 import { Button } from '@/components/ui/button'
-import { getUserData, isAuthenticated } from '@/lib/auth'
+import { useUser } from '@/lib/auth'
 import { cn } from '@/lib/utils'
 import { Review } from '@/types/review-section'
-import type { User } from '@/types/user'
 import { Pencil1Icon, TrashIcon } from '@radix-ui/react-icons'
-import { useEffect, useState } from 'react'
 
 export function EditDeleteReview({ review }: { review: Review }) {
-  // TODO: Fix cannot use useAuth because its parent is RSC
-  const [isAuth, setIsAuth] = useState(false)
-  const [user, setUser] = useState<User>()
-
-  useEffect(() => {
-    if (isAuthenticated()) {
-      setIsAuth(true)
-      const user = getUserData()
-      setUser(user)
-    }
-  }, [typeof window !== 'undefined'])
+  const user = useUser()
 
   function handleDeleteReview() {
     throw new Error('Function not implemented.')
@@ -27,7 +15,7 @@ export function EditDeleteReview({ review }: { review: Review }) {
     <div
       className={cn(
         'hidden',
-        isAuth && user?.user_id === review.user_id && 'flex',
+        user && user!.user_id === review.user_id && 'flex',
       )}
     >
       <Button size="icon" variant="ghost">
